- Deleted default course endpoints. - Updated extended course endpoints according to API. - Updated course endpoints tests. - Added doc.
