+++ src/api/endpoints_interfaces/IEndpointsCourseExtended.ts
@@ -5,4 +5,9 @@ import { ICourse } from "../../types/ICourse";
  */
 export interface IEndpointsCourseExtended {
     getFilteredCourses(getRequest: RequestInit, params?: ICourse): Promise<ICourse[]>;
+    getCourse(getRequest: RequestInit, id: number): Promise<ICourse>;
+    createCourse(postRequest: RequestInit): Promise<ICourse>;
+    updateCourse(putRequest: RequestInit): Promise<ICourse>;
+    patchCourse(putRequest: RequestInit): Promise<ICourse>;
+    deleteCourse(deleteRequest: RequestInit, id: number): void;
 }
