+++ src/components/CreateCourseComponent.tsx
@@ -7,13 +7,11 @@ import { validateCourseName } from "../helperScripts/validateCourseEntry";
 import { validateCourseDescription } from "../helperScripts/validateCourseEntry";
 import i18n from "../locales";
 import { RequestFactory } from "../api/requests/RequestFactory";
-import { EndpointsCourse } from "../api/endpoints/EndpointsCourse";
 import { loggerFactory } from "../../logger/LoggerConfig";
 import { CoursePublishState } from "../constants/CoursePublishState";
 import { EndpointsCourseExtended } from "../api/endpoints/EndpointsCourseExtended";
 
 const loggerService = loggerFactory.getLogger("service.CreateCourseComponent");
-const endpointsCourse: EndpointsCourse = new EndpointsCourse();
 const endpointsCourseExtended: EndpointsCourseExtended = new EndpointsCourseExtended();
 
 export const CreateCourseComponent: React.FC = () => {
@@ -86,7 +84,7 @@ export const CreateCourseComponent: React.FC = () => {
                         testID="courseIdInput"></TextInput>
                 </View>
                 <Pressable style={styles.styledButton}>
-                    <Button title={i18n.t("itrex.publishCourse")} onPress={updateCourse}></Button>
+                    <Button title={i18n.t("itrex.publishCourse")} onPress={patchCourse}></Button>
                 </Pressable>
                 <Pressable style={styles.styledButton}>
                     <Button title={i18n.t("itrex.getPublishedCourses")} onPress={getPublishedCourses}></Button>
@@ -127,18 +125,18 @@ export const CreateCourseComponent: React.FC = () => {
 
         loggerService.trace(`Creating course: name=${courseName}, startDate=${currentDate}.`);
         const postRequest: RequestInit = RequestFactory.createPostRequest(course);
-        endpointsCourse.createCourse(postRequest).then((data) => console.log(data));
+        endpointsCourseExtended.createCourse(postRequest).then((data) => console.log(data));
     }
 
     function getAllCourses(): void {
         loggerService.trace("Getting all courses.");
         const request: RequestInit = RequestFactory.createGetRequest();
-        endpointsCourse.getAllCourses(request).then((receivedCourses) => {
+        endpointsCourseExtended.getFilteredCourses(request).then((receivedCourses) => {
             setCourses(receivedCourses);
         });
     }
 
-    function updateCourse(): void {
+    function patchCourse(): void {
         loggerService.trace("Parsing ID string to ID number");
         const courseIdNumber: number = parseCourseId();
 
@@ -150,7 +148,7 @@ export const CreateCourseComponent: React.FC = () => {
 
         loggerService.trace(`Updating course: name=${courseName}, publishedState=${CoursePublishState.PUBLISHED}.`);
         const putRequest: RequestInit = RequestFactory.createPatchRequest(course);
-        endpointsCourse.updateCourse(putRequest).then((data) => console.log(data));
+        endpointsCourseExtended.patchCourse(putRequest).then((data) => console.log(data));
     }
 
     function parseCourseId(): number {
@@ -178,7 +176,7 @@ export const CreateCourseComponent: React.FC = () => {
         loggerService.trace("Parsing ID string to ID number");
         const courseIdNumber: number = parseCourseId();
 
-        endpointsCourse.deleteCourse(request, courseIdNumber);
+        endpointsCourseExtended.deleteCourse(request, courseIdNumber);
     }
 };
